pyproject.toml
[project]
name = "ai-by-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.60.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.1",
]
